+++ test_infra/Cargo.toml
@@ -9,6 +9,8 @@ dirs = "4.0.0"
 epoll = "4.3.1"
 libc = "0.2.132"
 once_cell = "1.14.0"
+serde = { version = "1.0.144", features = ["rc", "derive"] }
+serde_json = "1.0.85"
 ssh2 = { version = "0.9.3", features = ["vendored-openssl"] }
 vmm-sys-util = "0.10.0"
 wait-timeout = "0.2.0"
